requirements.txt
astroid==1.6.5
certifi==2018.4.16
chardet==3.0.4
defusedxml==0.5.0
Django==2.0.4
django-allauth==0.36.0
django-bootstrap3==10.0.1
django-cacheops==4.0.6
django-filter==1.1.0
django-icons==0.1.0
django-widget-tweaks==1.4.2
funcy==1.10.2
idna==2.6
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
oauthlib==2.0.7
Pillow==5.1.0
pkg-resources==0.0.0
pylint==1.9.2
pylint-django==0.11.1
pylint-plugin-utils==0.3
python3-openid==3.1.0
pytz==2018.4
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
urllib3==1.22
wrapt==1.10.11
